nit: move imports for resolve_kv_seq_len modeling_utils to _utils
