+++ QEfficient/transformers/models/granitemoe/modeling_granitemoe.py
@@ -28,7 +28,7 @@ from transformers.models.granitemoe.modeling_granitemoe import (
 
 from QEfficient.transformers.cache_utils import QEffDynamicCache
 from QEfficient.transformers.modeling_attn_mask_utils import _create_causal_mask
-from QEfficient.transformers.modeling_utils import resolve_kv_seq_len
+from QEfficient.utils._utils import resolve_kv_seq_len
 from QEfficient.utils.constants import MIN_MASKED_ATTENTION_VALUE
