+++ QEfficient/transformers/models/mixtral_moe/modeling_mixtral.py
@@ -32,7 +32,7 @@ from transformers.models.mixtral.modeling_mixtral import (
 
 from QEfficient.transformers.cache_utils import QEffDynamicCache
 from QEfficient.transformers.modeling_attn_mask_utils import _create_causal_mask
-from QEfficient.transformers.modeling_utils import resolve_kv_seq_len
+from QEfficient.utils._utils import resolve_kv_seq_len
 from QEfficient.utils.constants import MIN_MASKED_ATTENTION_VALUE
