+++ QEfficient/transformers/models/mllama/modeling_mllama.py
@@ -40,10 +40,9 @@ from QEfficient.transformers.modeling_utils import (
     _create_causal_mask,
     _prepare_aspect_ratio_attention_mask,
     _prepare_cross_attention_mask,
-    resolve_kv_seq_len,
 )
 from QEfficient.utils import constants
-from QEfficient.utils._utils import IOInfo
+from QEfficient.utils._utils import IOInfo, resolve_kv_seq_len
 from QEfficient.utils.constants import MIN_MASKED_ATTENTION_VALUE
 
 MAX_NUM_IMG = 1
