+++ QEfficient/transformers/models/qwen2/modeling_qwen2.py
@@ -30,7 +30,7 @@ from transformers.models.qwen2.modeling_qwen2 import (
 
 from QEfficient.transformers.cache_utils import QEffDynamicCache
 from QEfficient.transformers.modeling_attn_mask_utils import _create_causal_mask
-from QEfficient.transformers.modeling_utils import resolve_kv_seq_len
+from QEfficient.utils._utils import resolve_kv_seq_len
 from QEfficient.utils.constants import MIN_MASKED_ATTENTION_VALUE
